Tick Tac Toe rough idea
Based on research, there are code already available to use from 
Website 1: https://github.com/GokulVSD/tiCtactoe
Website 2: https://github.com/tindzk/GTK/blob/master/examples/tictactoe/tictactoe.c
Website 3: https://sudonull.com/post/148966-Playing-Tic-Tac-Toe-with-Python-and-GTK

so, for all the websites, especially the second one, there is something in common of them, which is they used button and grids to represents the  game grids. Therefore, I did some try out on creating the grids, however I think it is going to be challenging to integrate, them with the code that we have now. I will work on the code everyday to kind of try and improve it. That's all for the update so far!